src/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
SQLAlchemy==2.0.29
PyJWT==2.8.0
python-multipart==0.0.9